Smoke
I don't use anything other than GitHub to transfer copies of my code
between machines, and I want to go outside for a cigarette but continue
coding. So that's what this commit is for.

HttpServer is just tucking away the HTTP portion of the game server, and
I've begun writing some code for serving client files. Need to get this
working in tandem with PHP at some point.

Also, previously, I was calling a function with JavaScripts
document.onload as well as the body tag's onload attribute, by accident.
That's sorted now, and the frame rate is not doubled, and there are not
two players. :)

Next focus is building the lobby UI and getting it connected using
Socket.io