Progress
The HTML5 client now communicates fine with the Flash client using the
Node server alone. First time I've seen that happen, I've been dreaming
about it for weeks.

Next up is polishing the functionality of the client and server with
regards to keeping in sync with other users connected and updating
outifts as they're changed.

Very close to starting on building game worlds with HTML5!!