Ports
Set up now works locally and on dotcloud, with an identical codebase.
Visit http://gifted-hex72.dotcloud.com/ to see the client in action.

Legacy flash client will soon be included in the source to compare with.
The HTML5 client is a *lot* simpler, while retaining use of the same
protocol I wrote years ago. This means the Flash and HTML5 client are
now capable of talking to each other, and eventually, the intention is
to get the both simulating the game logic in exactly the same way.