Server > Client
Containing client-side code within the server directory. The client code
can be deployed completely independently to any server running PHP. If
the server can run PHP as well as Node, the client can be served
directly by the server. I've mainly found this useful for local testing.

In the event that the server is not running PHP, it will redirect the
browser to where the client is hosted. Currently these values are static
based on whether it's running locally or on dotcloud. Hopefully I'll be
able to store some form of configuration set up within the MySQL
database. It would be nice to have one seamless set up for the server
and client given a few key details.

This is also, in a way, an example of the server being authoritative,
making the client its little bitch.