Minor refactoring
Fixed that annoying client issue where, upon server restart and the
client automatically reconnects, it would rebind the message handler to
the same socket object and handle everything twice. All I needed to do
was move the message handler outside of the connect handler. Why on
Earth was it there in the first place?