MySQL connectivity
Also moved things around again. Client doesn't need to be in server,
only reason I did that was to get dotCloud working in the same way.

All I needed to do was nest the entire project in a src folder, which
actually saves me from keepin the dotcloud.yml in the repository, which
is nice.

MySQL works a breeze with this module.