Server-side maps nearly done
All that's left is saving to and loading from the database. That is not
much of a worry since all the flattening/expansion algorithms are sorted
and all that's required is performing queries and testing them
thoroughly.

I'm now going to jump ahead a bit and move work to the client; let the
user select from the available worlds on the server and join the lobby
of whichever, as well as see a preview of the map before/upon joining.

Time to get modular on the client-side for the first time in my life.