Tile drawing
So tiles are drawing fine in the canvas now, I just have to remember to
call this.initialize() whenever I extend an EaselJS container or library
asset (which isn't specified anywhere in the docs!) to make sure they
behave correctly and not in a static fashion.

Rather excited to see worlds drawing and before you know it I'll be onto
the physics engine and up to speed with the flash client in no time!
Then fresh progress with the game itself will resume.