Shared classes
Rough prototype of a shared class implemented with the server-side Map
class so that the client can use the same general code for handling
maps.

General idea is as follows:
Wrap entire definition in a function called init
If running on node, execute contents to assign the class to
module.exports
If running clientside with RequireJS, define the class as the init
function, seeing as RequireJS needs a function that returns the object
defining the module.

This could save me a lot of time in the future and help keep
functionality synchronised between server and client whenever updated.