Game loop, scrolling mechanics
Semi-efficient tile caching and redrawing mechanics implemented, it's a
bit messy but will bode well for saving memory.

The map remains stored as integers, but when a particular area is in
view, any tiles that get added to the world's map container replace the
integer in the map data with a reference to themselves. This is to allow
the engine to very quickly check whether the tile has been added to the
display list already or not.

I'm hoping this won't pose any problems, as the tile's toString method
returns what the integer value would be anyway. The map data is serving
its purpose without hogging memory; in the legacy client the entirety of
the data would be tile objects, but doing that in JavaScript is an
absolute memory hog. Doing it this way saves memory and keeps the
functionality in place.

I really hope this doesn't mess anything up in the future!