Physics engine
Getting this to work with the class structure I was using with AS3 was a
bit fiddly, but I have a blog post on it and it's actually quite simple,
so it would seem. Not fully tested at all, and not tested in any way on
the client yet, but it's getting there.

Involves further inheritance and, provided the client is having a good
day, shared classes. The shared logic will continually save me a lot of
time indeed.

Previous commit was forced because the GitHub app really struggles with
any files being moved or renamed. Not surprising I suppose.