World loop fixes
The correct modulo arithmetic that I've been looking for, for ages, and
almost thought of myself before finding it: ((n % M) + M) % M, as
applied in Map.js for the coordinate conversion. The x axis will now
loop perfectly, infinitely. This is going to be made optional for each
world/map, so that within some worlds can be smaller worlds (imagine the
inside of a house) that don't loop.