Live fixes
Small changes, some of which to fix the staging deployment on dotcloud.
Multiplayer is disabled for now. TCP handler also needs updating (if it
will at all be kept in the future for a Flash client).

Also, no more dedicated client service running PHP; the entire game can
run on an instance of node connected to a MySQL database.