Dynamic arm behaviour
The rules in place for dynamic arms are at the moment dependent on the
artwork used for the players. Hopefully in future I'll come up with a
way to standardise it a little so that the same code can work with any
art resources.

Mainly pushing because I'm about to start working on multiplayer
synchronisation, as well as renaming the server-side classes Room and
Rooms to World and Worlds to avoid unnecessary confusion with how the
client works. This functionality will not be shared with the client
however; sharing too much sharing client could easily lead to security
vulnerabilities and I want to be mindful of that as early as possible.