Minor config and boot changes for Dotcloud
These small changes allow the multiplayer server to work on dotcloud.
The changes to gifted.js (the programs entry point) look major from the
diffs, but it's actually just a rearrangement of requirejs calls. jQuery
and the global object are required first so that the address for
socket.io can be retrieved, then the rest of the program is loaded.

This is exactly the solution I wanted: one without having to regex in
the socket.io address whenever gifted.js is requested. Sweet.