Projectiles
A rough start has been made on projectile objects, which may or may not
be the super class for particle effects later on. The current focus is
on weapon projectiles.

Also includes minor changes to the physics engine and a crucial bug
fixed in the inheritance model of some classes (not reassigning the
constructor).