Debug features & minimap progress
Added debug option to display hitboxes. Debug output is now stored in
debugObj.out.

Reimplementing valueOf and toString in Tile was important for getting
map data to generate completely. Holding down M now displays an overlay
of the whole map. Very messy presentation with it at the moment, but
this is progress, not completion. The generation of the pixel data may
also need to be made asynchronous. The reason it is not done once and
left to be trusted is incase the world map changes throughout gameplay.

To aid this, World now has three containers. From bottom to top:

mapContainer (for tiles)
entityContainer (for every entity in the world)
overlay (for in-game ui)

All of these are handled by the world itself, and Entities now add
themselves as children to the entityContainer instead of the display
list of the world.

This is paving the way to having a nice environment for testing much
more advanced prodecural world generation, which I'm hoping to start on
soon.