Mini map, full map display, world gen progress
Mini map and full map implemented, rather hackily like many recent
commits (I'm starting to worry about time).

Implemented a very basic "miner" for the world generation algorithm and
created an underground spawn. The trick now is finding a way to create
miners that dig to the surface in a way that's always reachable by
players. Could prove a challenge.