Socket address, version number
The server only needs to respond with the port number, not the entire
address.

The socket server will always be running at the same address as the HTTP
server, so it's unnecessary to send the entire address from the server.
A simple, helpful change.

Version number has also been updated considering the progress with the
last few commits. v0.2 will be when all entities are fully synchronised
in multiplayer.