Item sync bug fixed, dynamic arms
Finally fixed that dreadful bug. It was simply because I was splicing
the projectile collection array on the server. I completely forgot that
this operation *does not* retain the key->value relationship. Currently
now just sets to false, may return to using delete() but that just sets
array values to undefined.

Also reworked the calculation of the angle used for the dynamic arm that
isn't in use so that other players see this as well, based on the main
aimAngle value. Comments were left behind to show workings and weird
things. Will make a commit immediately after this removing those.

Good progress! Let's get guns firing!