Bullet behaviour, pickup/drop, raytracing
Physics rules have been changed a tiny amount to make bullets behave
well.

The ordering of instructions for picking up and dropping items has been
modified to make checking whether to send an update for item
creation/destruction is necessary.

Bresenham's line algorithm has been implemented to retrieve an array of
x,y's representing a line in the direction an Entity is travelling. This
is for the sake of raytracing, and will be used on the server side for
precise collision detection between Bullet and Character objects.