In-game menus, respawning, socket flow fixes
An in-game menu has been added. Accessible by pressing ESC. Upon death
users are presented with a respawn button. This works with multiplayer
too.

UI fix for map placement when quitting/resetting the game, very hacky
but it works.

Socket flow has been improved so that upon an attempt to connect when
the server is down, the user is taken back to the main menu, and upon
connecting after such a failed connect, everything works normally. No
error messages for this yet.